+++ src/test/java/ar/edu/unlam/tallerweb1/delivery/ControladorMascotaTest.java
@@ -63,8 +63,8 @@ public class ControladorMascotaTest {
     }
     @Test
     public void buscarMascotaPorIdUsuarioIncorrecto(){
-        Long idUsuario = 1l;
-        Long idUsuarioIncorrecto = 2l;
+        int idUsuario = 1;
+        int idUsuarioIncorrecto = 2;
         dadoQueExisteMascota(idUsuario);
         ModelAndView mav = this.controladorMascotas.getMascotaPorIdUsuario(idUsuarioIncorrecto);
         assertThat(mav.getModel().get("mascotas").toString()).isEqualTo("[]");
@@ -97,7 +97,7 @@ public class ControladorMascotaTest {
         assertThat(result.getBody()).isNull();
     }
 
-    private void dadoQueExisteMascota(Long idUsuario) {
+    private void dadoQueExisteMascota(int idUsuario) {
         List<Mascota> mascotas = new ArrayList<>();
         Mascota mascota = new Mascota();
         mascota.setIdUsuario(idUsuario);
@@ -149,7 +149,7 @@ public class ControladorMascotaTest {
     }
     @Test
     public void buscarMascotaPorIdUsuario(){
-        Long idUsuario = 1l;
+        int idUsuario = 1;
         dadoQueExisteMascota(idUsuario);
         ModelAndView mav = this.controladorMascotas.getMascotaPorIdUsuario(idUsuario);
         assertThat(mav.getModel().get("mascotas")).isNotNull();
